chore: move PaymentMethodParams to its own module
