+++ packages/theme/composables/usePaymentProvider/PaymentMethodParams.ts
@@ -1,0 +1,8 @@
+import type { PaymentMethodInput } from '~/modules/GraphQL/types';
+
+interface PaymentMethodParams {
+  cart_id: string;
+  payment_method: PaymentMethodInput;
+}
+
+export default PaymentMethodParams;
