+++ packages/theme/composables/usePaymentProvider/usePaymentProvider.ts
@@ -1,6 +1,1 @@
-import type { PaymentMethodInput } from '~/modules/GraphQL/types';
-
-export interface PaymentMethodParams {
-  cart_id: string;
-  payment_method: PaymentMethodInput;
-}
+export {};
